// src/generated/prisma/models/Card.ts
export type CardModel = runtime.Types.Result.DefaultSelection<Prisma.$CardPayload>
// ...
export type $CardPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Card"
  objects: {
    deckCards: Prisma.$DeckCardPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: number
    name: string
    hp: number
    attack: number
    type: $Enums.PokemonType
    pokedexNumber: number
    imgUrl: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["card"]>
  composites: {}
}